# docs/swagger.yaml
definitions:
  focalboard-tool_library_render.JSON:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/focalboard/boards/single:
    get:
      parameters:
      - description: Board Id
        in: query
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/focalboard-tool_library_render.JSON'
      summary: Get a single board by its ID.
...
swagger: "2.0"